minor zipkin refactor
